core: Introduce MR with dynamic keys
Memory registration consists of two parts: map/pin the memory for local
access and export with a key for remote access. The first part is usually
heavyweight and requries kernel involvement. The second part is less
expensive and can be further separated into key allocation and key
assignment. Key allocation may needs kernel involvement, but key assignment
can be done in user space. This leads to the concept of MR with dynamic key.

Dynamic key allows an MR be exported as different keys for different target,
with the option to expose different address range and access rights, all
without the overhead of registering the MR multiple times in the standard
way.

A new flag FI_MR_DYNAMIC_KEY is introduces to be used with memory registration
calls. This flag also serves as a secondary capbility.

New API functions are added to allocate / assign / revoke keys.

Another flag FI_MR_SINGLE_USE is added for single use key assignment.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>